+++ src/main/java/com/lms/lmsapi/repository/UserTypeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.lms.lmsapi.entity.UserType;
 
 @Repository
-public interface UserTypeRepository extends JpaRepository<UserType, Integer>{
+public interface UserTypeRepository extends JpaRepository<UserType, Long>{
     
 }
